requirements.txt
asgiref==3.6.0
attrs==23.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
cryptography==40.0.2
defusedxml==0.7.1
Django==3.2
django-filter==23.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
flake8==6.0.0
idna==3.4
iniconfig==2.0.0
isort==5.12.0
mccabe==0.7.0
oauthlib==3.2.2
packaging==23.1
pluggy==0.13.1
py==1.11.0
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
requests==2.26.0
requests-oauthlib==1.3.1
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
toml==0.10.2
urllib3==1.26.15
